Feat: :sparkles: add Repository, Controller, Service, Exceptions, config and auth

+++ src/main/java/com/Campusland/ProyectoSpringBoot_CorpusEnrique/dto/response/AuditoriaResponse.java
@@ -4,14 +4,13 @@ import com.Campusland.ProyectoSpringBoot_CorpusEnrique.model.Auditoria.TipoOpera
 import java.time.LocalDateTime;
 import java.util.Map;
 
-public record AuditoriaResponse ( Integer idAuditoria,
+public record AuditoriaResponse ( Long idAuditoria,
         UsuarioResponse usuario,
         String entidadAfectada,
-        Integer registroId,
+        Long registroId,
         TipoOperacion tipoOperacion,
         Map<String, Object> valoresAnteriores,
         Map<String, Object> valoresNuevos,
         LocalDateTime fechaHora){
-
 
 }
