+++ src/main/java/com/Campusland/ProyectoSpringBoot_CorpusEnrique/dto/response/BodegaResponse.java
@@ -2,7 +2,7 @@ package com.Campusland.ProyectoSpringBoot_CorpusEnrique.dto.response;
 
 import com.Campusland.ProyectoSpringBoot_CorpusEnrique.model.Bodega.Estado;
 
-public record BodegaResponse (  Integer idBodega,
+public record BodegaResponse (  Long idBodega,
         String nombre,
         String direccion,
         Integer capacidad,
