+++ src/main/java/com/Campusland/ProyectoSpringBoot_CorpusEnrique/dto/response/CategoriaResponse.java
@@ -1,7 +1,7 @@
 package com.Campusland.ProyectoSpringBoot_CorpusEnrique.dto.response;
 
 
-public record CategoriaResponse ( Integer idCategoria,
+public record CategoriaResponse ( Long idCategoria,
         String nombre){
