+++ src/main/java/com/Campusland/ProyectoSpringBoot_CorpusEnrique/dto/response/CiudadResponse.java
@@ -1,7 +1,7 @@
 package com.Campusland.ProyectoSpringBoot_CorpusEnrique.dto.response;
 
 
-public record CiudadResponse(     Integer idCiudad,
+public record CiudadResponse(     Long idCiudad,
                                  String nombre,
                                  DepartamentoResponse departamento) {
