+++ src/main/java/com/Campusland/ProyectoSpringBoot_CorpusEnrique/dto/response/PersonaResponse.java
@@ -3,7 +3,7 @@ package com.Campusland.ProyectoSpringBoot_CorpusEnrique.dto.response;
 import com.Campusland.ProyectoSpringBoot_CorpusEnrique.model.Persona.TipoDocumento;
 
 public record PersonaResponse(
-    Integer idPersona,
+    Long idPersona,
     String nombre,
     String apellido,
     TipoDocumento tipoDocumento,
