+++ src/main/java/com/Campusland/ProyectoSpringBoot_CorpusEnrique/dto/response/ProductoResponse.java
@@ -1,7 +1,7 @@
 package com.Campusland.ProyectoSpringBoot_CorpusEnrique.dto.response;
 
 public record ProductoResponse(
-    Integer idProducto,
+    Long idProducto,
     String nombre,
     String codigo,
     String unidadMedida,
