+++ src/main/java/com/Campusland/ProyectoSpringBoot_CorpusEnrique/dto/response/UsuarioResponse.java
@@ -3,7 +3,7 @@ package com.Campusland.ProyectoSpringBoot_CorpusEnrique.dto.response;
 import com.Campusland.ProyectoSpringBoot_CorpusEnrique.model.Usuario.Estado;
 
 public record UsuarioResponse(
-    Integer idUsuario,
+    Long idUsuario,
     String username,
     Estado estado,
     PersonaResponse persona,
